TASK: All to execute a callback with or without indexing disabled
This change introduce two new methods to enable or not the indexing during
a request. This can be useful during node import where the indexing of a
parent node is dependent on the children node, in the case the indexing of
the parent node can be done only after the children node are imported.