TASK: Run latest migrations
* TASK: Apply migration TYPO3.FluidAdaptor-20161130112935

Adjusts code to package renaming from "TYPO3.Fluid" to
"Neos.FluidAdaptor".

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* TASK: Apply migration Neos.Fusion-20161201202543

Moves fusion files from old path ``Resources/Private/TypoScript/`` to
the new path

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* TASK: Apply migration Neos.Neos-20161201222211

Migrate namespaces for fusion core implementation and helper classes

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* TASK: Apply migration Neos.Fusion-20161202215034

Migrate name for the Fusion content cache to Neos_Fusion_Content

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* TASK: Apply migration Neos.Fusion-20161219092345

Migrate name for the Fusion cache to Neos_Neos_Fusion

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* TASK: Apply migration Neos.ContentRepository-20161219093512

Migrate name for the CR node type configuration

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* TASK: Apply migration Neos.Media-20161219094126

Migrate name for the media image size cache

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* TASK: Apply migration Neos.Neos-20161219094403

Migrate several cache keys for the Neos.Neos package

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* TASK: Apply migration Neos.Neos-20161219122512

Migrate usages of TypoScriptService to FusionService

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* TASK: Apply migration Neos.Fusion-20161219130100

Migrate name for the TypoScriptView to FusionView

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* TASK: Apply migration Neos.Neos-20161220163741

Migrate usages of the Settings path Neos.Neos.typoScript to
Neos.Neos.fusion

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* TASK: Apply migration Neos.Neos-20170115114620

Migrate usages of the Settings path
Neos.Flow.security.authentication.providers.Typo3BackendProvider to
Neos.Flow.security.authentication.providers[Neos.Neos:Backend]

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* TASK: Apply migration Neos.Fusion-20170120013047

Migrate name for the TypoScriptView to FusionView

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* TASK: Apply migration Neos.Flow-20170125103800

Migrate usages of the path [TYPO3][Flow][Security][Authentication] to
[Neos][Flow][Security][Authentication]

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* TASK: Apply migration Neos.Seo-20170127154600

Migrate to new namespace

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.

* TASK: Apply migration Neos.Flow-20170127183102

Migrate bootstep names.

Note: This migration did not produce any changes, so the commit simply
marks the migration as applied. This makes sure it will not be applied
again.