[BUGFIX] Fix version constraints for flow version 5
The latest version 3.0.8 changed the version constraints, so that the package version was not available anymore for the flow 5.x users. The versions before were available and as the version 3.0.8 makes the package PHP 7.4 compatible the users of Neos 4.3 and flow 5.3 wants to use the version, of course.

This change just brings back the compatibility for flow 5. With higher versions, the package is not compatible as the interfaces for the QueryBuilder for instance changed, and the constraint was more a result of a mistake.

# Conflicts:
#	composer.json